ContactForm ContactItem Reducer ContactState
we make current object null where we are doing in reducer and get the contact as an object and we make the current null if we do not want edit something. In ContactForm we use UseEffect hook if we have data in curent then it  wil assign in   setContact. In ContactItem we are deleteing the specific User via id we are doing filter.